Gather http routes in an exclusive module
