+++ cast/app.py
@@ -13,16 +13,11 @@ from flask import (
 import sys
 
 import flask_socketio
+from flask_jwt_extended import JWTManager
 
-from flask_jwt_extended import (
-    JWTManager,
-    jwt_required,
-    create_access_token,
-    set_access_cookies,
-)
 
-from werkzeug.exceptions import BadRequest
 import json
+from .routes import http
 from .logger import Logging
 import pty
 import os
@@ -43,6 +38,8 @@ app.config["current_session"] = None
 app.config["sessions"] = {}
 app.config["log_file"] = r"log_data/"
 app.config["JWT_TOKEN_LOCATION"] = ["cookies"]
+
+app.register_blueprint(http, url_prefix=r"")
 
 
 def random_string(string_length=10):
@@ -96,35 +93,6 @@ def unauthorized_response(callback):
 @jwt.invalid_token_loader
 def invalid_token(callback):
     return redirect("/")
-
-
-@app.errorhandler(404)
-def page_not_found(error):
-    return render_template("404.html"), 404
-
-
-@app.route("/cast")
-@jwt_required
-def index():
-    log = Logging(app.config["current_session"])
-    log.make_log_folder()
-    return render_template("index.html")
-
-
-@app.route("/", methods=["GET", "POST"])
-def login():
-    if request.method == "GET":
-        return render_template("login.html")
-
-    elif request.method == "POST":
-        data = request.json
-        if data["password"] == app.config["passwd"]:
-            access_token = create_access_token(identity=os.getenv("JOB_ID"))
-            resp = jsonify({"login": True})
-            set_access_cookies(resp, access_token)
-            return resp
-        else:
-            return json.dumps(request.get_json()), 401
 
 
 @socketio.on("new-session", namespace="/cast")
@@ -240,17 +208,6 @@ def client_input(data):
             else:
                 log.write_log(data["input"])
                 os.write(file_desc, data["input"].encode())
-
-
-# This is the route handler for DOWNLOADING the log file. Maybe a bit buggy. Please report if found
-@app.route("/download/<string:file_path>")
-def download(file_path):
-    try:
-        return send_from_directory(
-            app.config["log_file"], filename=file_path, as_attachment=True
-        )
-    except BadRequest:
-        return 404
 
 
 def create_parser():
